Projeto usa tempo entre updates como referencia para suas contas e nao quantidade de updates. --repetitions trocado por --duration (-T). main.c usa objectTable agora. (personTable esta obsoleta, mas é bom deixar ali para pegar codigo) Deve ter mais coisas, mas eu sempre acabo tendo que mexer no projeto inteiro para mudar algo, pqp
git-svn-id: svn://svn/henriquelima/labprog@155 c5a4425f-b1d8-4b97-91a7-095824198b80

+++ lib/class.c
@@ -8,7 +8,7 @@
     
 static Class classList[NUM_CLASS];
 void classAdd(objectType type,
-        void (*updateFunc)(object, int, int),
+        void (*updateFunc)(object, int, double),
         void (*removeFunc)(object),
         void (*collideFunc)(object, object),
         void (*boundsFunc) (object),
